EPBR-5340 raise errors and print notifaction status to console

+++ lib/use_case/export_heat_pump_by_property_type.rb
@@ -10,10 +10,14 @@ module UseCase
 
     def execute(start_date:, end_date:, template_id:, email_address:)
       file_name = @file_gateway.file_name
-      @raw_data = @export_gateway.fetch_by_property_type(start_date:, end_date:)
-      @file_gateway.save_csv(@raw_data)
+      raw_data = @export_gateway.fetch_by_property_type(start_date:, end_date:)
+
+      raise Boundary::NoData, "heat pump data" unless raw_data.any?
+
+      @file_gateway.save_csv(raw_data)
       @notify_gateway.send_email(template_id:, file_name:, email_address:)
-      File.delete(file_name)
+      File.delete(file_name) if File.exists?(file_name)
+      @notify_gateway.check_email_status
     end
   end
 end
